♻️ [Refactoring]: reorganize deployment module layout
- Move ComponentDeployment struct and basic accessors from executor.rs back to deployment.rs root so the module root owns the type and submodules hold only behavior
- Rename bash_escape.rs / bash_escape_test.rs to bash.rs / bash_test.rs since the file holds bash utilities, not just escaping
