+++ src/component/deployment/bash_test.rs
